pyproject.toml
[project]
name = "rag-pipeline"
version = "0.1.0"
description = "Multi-modal RAG pipeline for long-document understanding"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "black>=25.11.0",
    "colpali-engine",
    "faiss-cpu>=1.13.0",
    "google-generativeai>=0.8.5",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.0.1",
    "langchain-openai>=1.0.3",
    "langchain-text-splitters>=1.0.0",
    "openai>=2.8.1",
    "pymupdf>=1.26.6",
    "pytesseract>=0.3.13",
    "pyyaml>=6.0.3",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.2",
    "torch==2.8.0",
    "transformers>=4.57.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["data*", ".venv*", "*.egg-info"]

[tool.uv.sources]
colpali-engine = { git = "https://github.com/illuin-tech/colpali" }
